fix: batch enum-enum bitwise OR fixes for Clang 16

+++ submodules/LegacyComponents/Sources/GPUImageFramebuffer.m
@@ -385,7 +385,7 @@ void dataProviderUnlockCallback (void *info, __unused const void *data, __unused
         }
         else
         {
-            cgImageFromBytes = CGImageCreate((int)_size.width, (int)_size.height, 8, 32, 4 * (int)_size.width, defaultRGBColorSpace, kCGBitmapByteOrderDefault | kCGImageAlphaPremultipliedLast, dataProvider, NULL, NO, kCGRenderingIntentDefault);
+            cgImageFromBytes = CGImageCreate((int)_size.width, (int)_size.height, 8, 32, 4 * (int)_size.width, defaultRGBColorSpace, (uint32_t)kCGBitmapByteOrderDefault | (uint32_t)kCGImageAlphaPremultipliedLast, dataProvider, NULL, NO, kCGRenderingIntentDefault);
         }
         
         // Capture image with current device orientation
